tests/wpt/web-platform-tests/css/css-transforms/transform_translate_max.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate with maximum translation value</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform sets translate(INFINITE, INFINITE),
                                 transform property returns initial value.">
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(INFINITE, INFINITE)";
        var value = window.getComputedStyle(document.getElementById("test")).getPropertyValue("transform");
        assert_equals(value, "none", "transform property value should be none");
    }, "transform_translate_max");
    </script>
</body>
</html>
